+++ packages/@cdktf/cli-core/src/lib/models/terraform-cli.ts
@@ -150,12 +150,22 @@ export class TerraformCli implements Terraform {
     );
   }
 
-  public async plan(
-    destroy = false,
-    refreshOnly = false,
-    parallelism = -1,
-    noColor = false
-  ): Promise<void> {
+  public async plan(opts: {
+    destroy: boolean;
+    refreshOnly?: boolean;
+    parallelism?: number;
+    vars?: string[];
+    varFiles?: string[];
+    noColor?: boolean;
+  }): Promise<void> {
+    const {
+      destroy = false,
+      refreshOnly = false,
+      parallelism = -1,
+      vars = [],
+      varFiles = [],
+      noColor = false,
+    } = opts;
     const options = ["plan", "-input=false"];
 
     if (destroy) {
@@ -170,6 +180,14 @@ export class TerraformCli implements Terraform {
     if (noColor) {
       options.push("-no-color");
     }
+
+    vars.forEach((v) => options.push(`-var='${v}'`));
+    varFiles.forEach((v) => options.push(`-var-file='${v}'`));
+
+    logger.debug(
+      `Executing ${terraformBinaryName} ${options.join(" ")} in ${this.workdir}`
+    );
+
     await this.setUserAgent();
 
     await exec(
@@ -192,6 +210,8 @@ export class TerraformCli implements Terraform {
       noColor = false,
       parallelism = -1,
       extraOptions = [],
+      vars = [],
+      varFiles = [],
     },
     callback: (state: TerraformDeployState) => void
   ): Promise<{ cancelled: boolean }> {
@@ -204,6 +224,8 @@ export class TerraformCli implements Terraform {
       autoApprove,
       parallelism,
       extraOptions,
+      vars,
+      varFiles,
     });
     return this.handleService("deploy", service, callback);
   }
@@ -214,6 +236,8 @@ export class TerraformCli implements Terraform {
       parallelism = -1,
       noColor = false,
       extraOptions = [],
+      vars = [],
+      varFiles = [],
     },
     callback: (state: TerraformDeployState) => void
   ): Promise<{ cancelled: boolean }> {
@@ -225,6 +249,8 @@ export class TerraformCli implements Terraform {
       parallelism,
       noColor,
       extraOptions,
+      vars,
+      varFiles,
     });
     return this.handleService("destroy", service, callback);
   }