+++ packages/cdktf-cli/src/bin/cmds/deploy.ts
@@ -84,6 +84,18 @@ class Command extends BaseCommand {
         required: false,
         desc: "Disables terminal formatting sequences in the output.",
       })
+      .option("var", {
+        type: "array",
+        default: [],
+        required: false,
+        desc: "Set a value for one of the input variables in the stack or stacks to apply. Use this option more than once to set more than one variable.",
+      })
+      .option("var-file", {
+        type: "array",
+        default: [],
+        required: false,
+        desc: "Load variable values from the given file, in addition to the default files terraform.tfvars and *.auto.tfvars. Use this option more than once to include more than one variables file.",
+      })
       .showHelpOnFail(true);
 
   public async handleCommand(argv: any) {