+++ packages/cdktf/lib/terraform-hcl-module.ts
@@ -1,10 +1,10 @@
 // Copyright (c) HashiCorp, Inc
 // SPDX-License-Identifier: MPL-2.0
 import { Construct } from "constructs";
-import { TerraformModuleOptions, TerraformModule } from "./terraform-module";
+import { TerraformModuleConfig, TerraformModule } from "./terraform-module";
 import { Token, IResolvable } from "./tokens";
 
-export interface TerraformHclModuleOptions extends TerraformModuleOptions {
+export interface TerraformHclModuleConfig extends TerraformModuleConfig {
   readonly variables?: { [key: string]: any };
 }
 
@@ -12,11 +12,7 @@ export interface TerraformHclModuleOptions extends TerraformModuleOptions {
 export class TerraformHclModule extends TerraformModule {
   private _variables?: { [key: string]: any };
 
-  constructor(
-    scope: Construct,
-    id: string,
-    options: TerraformHclModuleOptions
-  ) {
+  constructor(scope: Construct, id: string, options: TerraformHclModuleConfig) {
     super(scope, id, options);
 
     this._variables = options.variables;